Filled in topic skeleton.
Added considerably to the Data Collection section, and wrote a large amount of material for the Data Modeling section. The Training and Testing section needs a lot more work, including the inclusion of images that illustrate the classifier parameter selection process and choice of 10-fold cross-validation when calculating the main classifier performance metric (ROC AUC). The Data Exploration and Conclusion sections are still in need of fleshing out.